requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
idna==3.10
mysqlclient==2.2.7
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
